chore: add root conftest.py with shared test fixtures
Add tests/conftest.py with:
- Generic sample_dataset fixture (shared by text_concat, flow, and mcp_agent tests)
- Unified MockMessage class (combines LLMChatBlock dict-like and MCPAgentBlock
  model_dump interfaces)
- mock_message_class fixture for accessing MockMessage in test fixtures

Refactor test files to use shared fixtures:
- test_text_concat.py: use shared sample_dataset
- tests/flow/conftest.py: use shared sample_dataset
- test_mcp_agent_block.py: use shared sample_dataset and mock_message_class
- test_llm_chat_block.py: use mock_message_class

Clean up tech-debt-tracker.md:
- Remove resolved items (isort/pylintrc, __pycache__, .env.example, conftest)
- Keep remaining open items
